+++ back/imfine/src/main/java/com/idle/imfine/data/dto/bamboo/response/ResponseBambooDetailDto.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Builder
 public class ResponseBambooDetailDto {
-    private int bambooId;
+    private Long bambooId;
     private String content;
     private LocalDateTime createdDate;
     private int likeCount;
